agent/pyproject.toml
[project]
name = "dexter"
version = "0.1.0"
description = "AI-powered cybersecurity assistant with RAG capabilities"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "PyPDF2>=3.0.0",
    "langchain>=0.1.0",
    "langchain-core>=0.1.0",
    "langchain-community>=0.0.10",
    "langchain-openai>=0.0.5",
    "langchain-google-genai>=0.0.6",
    "python-dotenv>=1.0.0",
    "google-generativeai>=0.3.0",
    "faiss-cpu>=1.7.4",
    "langgraph>=0.0.15",
    "typing-extensions>=4.8.0",
    "openai>=1.12.0"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["core", "utils", "config"]

[dependency-groups]
dev = [
    "ruff>=0.9.5",
]
